+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Create: Mobile Packages - v0.7.7 - 1.21.1 - unreleased
 - Fix discord Curseforge link in ci
 - Fix BeePortBlockEntity.onChunkUnloaded() synchronously loading a chunk (#350)
 - Fix Owners are now members of a network by default (#346)
+- Fix Playernetworks Screen now showing network connection status: Member / Owner (#346)
 
 ------------------------------------------------------
 Create: Mobile Packages - v0.7.6 - 1.21.1 - 08.05.2026